refac: rename communication to websocket
