+++ watcher/fileHandler.go
@@ -3,10 +3,10 @@ package watcher
 import (
 	"log"
 
-	"github.com/CTNOriginals/BitburnerGoFilesync/communication"
-	"github.com/CTNOriginals/BitburnerGoFilesync/communication/constructor"
-	"github.com/CTNOriginals/BitburnerGoFilesync/communication/definitions"
 	"github.com/CTNOriginals/BitburnerGoFilesync/utils"
+	"github.com/CTNOriginals/BitburnerGoFilesync/websocket"
+	"github.com/CTNOriginals/BitburnerGoFilesync/websocket/constructor"
+	"github.com/CTNOriginals/BitburnerGoFilesync/websocket/definitions"
 )
 
 var FileEventHandlerMap = MFileEventHandler{
@@ -27,7 +27,7 @@ var FileEventHandlerMap = MFileEventHandler{
 
 		var relPath = file.RelativePath()
 
-		communication.SendRequest(definitions.DeleteFile, func(message *constructor.Message) {
+		websocket.SendRequest(definitions.DeleteFile, func(message *constructor.Message) {
 			if message.IsError {
 				log.Printf("\nwatcher.OnFileDelete: Unable to delete file: %s\nResponse: %v\n\n", relPath, message.Response)
 			}
@@ -41,7 +41,7 @@ func PushFile(file *FileInfo) {
 	var relPath = file.RelativePath()
 	var content = utils.SanitizeFileContent(utils.GetFileContentByPath(relPath))
 
-	communication.SendRequest(definitions.PushFile, func(message *constructor.Message) {
+	websocket.SendRequest(definitions.PushFile, func(message *constructor.Message) {
 		if message.IsError {
 			log.Printf("\nwatcher.PushFile: Unable to push file content: %s\nResponse: %v\n\n", relPath, message.Response)
 		}
