+++ websocket/constructor/message.go
@@ -1,7 +1,7 @@
 package constructor
 
 import (
-	"github.com/CTNOriginals/BitburnerGoFilesync/communication/definitions"
+	"github.com/CTNOriginals/BitburnerGoFilesync/websocket/definitions"
 
 	ctnstruct "github.com/CTNOriginals/CTNGoUtils/v2/struct"
 )
