+++ websocket/definitions/definition.go
