+++ websocket/definitions/definitionMap.go
