+++ websocket/tests.go
@@ -1,12 +1,10 @@
-package communication
+package websocket
 
 import (
-	"log"
-
 	"github.com/CTNOriginals/BitburnerGoFilesync/cli"
-	"github.com/CTNOriginals/BitburnerGoFilesync/communication/constructor"
-	"github.com/CTNOriginals/BitburnerGoFilesync/communication/definitions"
 	"github.com/CTNOriginals/BitburnerGoFilesync/config"
+	"github.com/CTNOriginals/BitburnerGoFilesync/websocket/constructor"
+	"github.com/CTNOriginals/BitburnerGoFilesync/websocket/definitions"
 )
 
 func TestCommunication() {
@@ -21,8 +19,6 @@ func TestCommunication() {
 	SendRequest(definitions.GetAllFiles, func(message *constructor.Message) {
 		// log.Printf("\nGetAllFiles OnResponse: {\n%s\n}\n", ctnstring.Indent(message.String(), 2, " "))
 	}, "home")
-
-	log.Println("heaslkdjef")
 
 	cli.CommandWatcher()
 }
